+++ src/containers/authGuard/initLoader/InitLoader.module.scss
@@ -11,7 +11,7 @@
   width: 200px;
   height: 200px;
 
-  background-color: rgb(56 39 102 / 30%);
+  background-color: rgb(19 8 35 / 30%);
   border-radius: 50%;
 
   animation: animate 1s linear infinite;
@@ -62,7 +62,7 @@
   span {
     position: absolute;
     inset: 20px;
-    background: var(--Dark-900);
+    background: var(--Dark-700);
     border-radius: 50%;
   }
 }
