+++ src/containers/index.ts
@@ -1,1 +1,3 @@
 export { AuthGuard } from './authGuard/AuthGuard'
+export { getLayout } from './layouts/layout/Layout'
+export { getLayoutWithNav } from './layouts/layoutWithNav/LayoutWithNav'
