+++ src/containers/layouts/components/navBar/NavBar.module.scss
