+++ src/containers/layouts/components/navBar/navBarItems/NavBarItems.module.scss
