+++ src/pages/github/Github.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from 'react'
 import { toast } from 'react-toastify'
 
+import { getLayout } from '@/containers'
 import { PATH, useRouterLocaleDefinition } from '@/shared'
-import { getBaseLayout } from '@/widgets'
 import { useRouter } from 'next/router'
 
 function GithubPage() {
@@ -22,5 +22,5 @@ function GithubPage() {
   }, [accessToken])
 }
 
-GithubPage.getLayout = getBaseLayout
+GithubPage.getLayout = getLayout
 export default GithubPage
