+++ src/widgets/index.ts
@@ -2,10 +2,6 @@
 export { NewPassword } from './auth/newPassword/NewPassword'
 export { ResendLink } from './auth/resendLink/ResendLink'
 
-// ============================== [ layouts ] ======================================
-export { getBaseLayout } from './layouts/BaseLayout'
-export { getLayoutWithNav } from './layouts/LayoutWithNav'
-
 export { Devices } from './profile/settings/devices/Devices'
 // ============================== [ profile ] ======================================
 export { GeneralInfo } from './profile/settings/generalInfo/GeneralInfo'
